OmniSwitch: locked apps support
allows tagging apps as "locked" which means they will not
be killed on actions kill all and kill other

explicitly killing a locked task will remove it from
the locked list and kill it in one step.

the locked apps are stored as a list
of package names. this list is updated on package changes
like uninstall and will remove uninstalled apps then

also add an option to sort locked apps on top
of the recent list

Change-Id: I08ff635657aa55db47a28c53a0f474e82689e788
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>